@@ -1,6 +1,10 @@
 {
   "name": "mm-admin-app",
-  "version": "0.0.0",
+  "version": "0.0.1",
+  "author": "Our Mindful Menu",
+  "description": "Our Mindful Menu - Admin App",
+  "homepage": "https://www.ourmindfulmenu.com/",
+  "private": true,
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -10,33 +14,47 @@
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@firebase/app": "^0.1.10",
+    "@angular/animations": "^5.2.10",
+    "@angular/common": "^5.2.10",
+    "@angular/compiler": "^5.2.10",
+    "@angular/core": "^5.2.10",
+    "@angular/forms": "^5.2.10",
+    "@angular/http": "^5.2.10",
+    "@angular/platform-browser": "^5.2.10",
+    "@angular/platform-browser-dynamic": "^5.2.10",
+    "@angular/router": "^5.2.10",
     "@ng-bootstrap/ng-bootstrap": "^1.1.2",
+    "@uirouter/angular": "^1.0.1",
+    "angular-async-local-storage": "^5.0.1",
+    "angular-event-service": "^1.2.0",
     "bootstrap": "^4.1.0",
+    "class-transformer": "^0.1.9",
     "core-js": "^2.4.1",
-    "firebase": "^4.13.0",
+    "firebase": "^4.13.1",
+    "glyphicons-only-bootstrap": "^1.0.1",
+    "jquery": "^3.3.1",
+    "lodash": "^4.17.10",
+    "moment": "^2.22.1",
+    "moment-timezone": "^0.5.16",
+    "ng4-validators": "^5.1.0",
+    "ngx-loading": "^1.0.14",
+    "ngx-pipes": "^2.1.2",
+    "ngx-toastr": "^8.4.0",
+    "popper.js": "^1.14.3",
     "rxjs": "^5.5.6",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler-cli": "^5.2.10",
+    "@angular/language-service": "^5.2.10",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
+    "@types/jquery": "^3.3.1",
+    "@types/lodash": "^4.14.108",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -47,6 +65,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "^2.6.2"
   }
 }
