@@ -1,12 +1,14 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/ApplicationModule';
-import { environment } from './environments/environment';
+import { ApplicationModule } from './app/ApplicationModule';
+import { environment } from './environment/environment';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule)
-  .catch(err => console.log(err));
+platformBrowserDynamic().bootstrapModule(ApplicationModule)
+  .catch((error) => {
+      console.log(error);
+  });
